+++ common/src/main/java/traben/entity_texture_features/features/property_reading/properties/optifine_properties/TimeOfDayProperty.java
@@ -2,12 +2,13 @@ package traben.entity_texture_features.features.property_reading.properties.opti
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import traben.entity_texture_features.config.ETFConfig;
 import traben.entity_texture_features.features.property_reading.properties.generic_properties.LongRangeFromStringArrayProperty;
 import traben.entity_texture_features.utils.ETFEntity;
 
 import java.util.Properties;
 
-import static traben.entity_texture_features.ETFClientCommon.ETFConfigData;
+
 
 public class TimeOfDayProperty extends LongRangeFromStringArrayProperty {
 
@@ -37,7 +38,7 @@ public class TimeOfDayProperty extends LongRangeFromStringArrayProperty {
 
     @Override
     public boolean isPropertyUpdatable() {
-        return !ETFConfigData.restrictDayTime;
+        return !ETFConfig.getInstance().restrictDayTime;
     }
 
     @Override