+++ common/src/main/java/traben/entity_texture_features/mixin/entity/misc/MixinGlowSquidEntity.java
@@ -7,7 +7,7 @@ import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
-import traben.entity_texture_features.ETFClientCommon;
+import traben.entity_texture_features.config.ETFConfig;
 import traben.entity_texture_features.features.ETFManager;
 
 @Mixin(GlowSquidEntity.class)
@@ -25,7 +25,7 @@ public abstract class MixinGlowSquidEntity extends SquidEntity {
             index = 2
     )
     private double mixin(double x) {
-        if (ETFClientCommon.ETFConfigData.enableCustomTextures
+        if (ETFConfig.getInstance().enableCustomTextures
                 && ETFManager.getInstance().ENTITY_TYPE_IGNORE_PARTICLES.contains(this.getType())) {
             return -500;
         }